Milestone 2 Doc
Please take the time to carefully look through this.  I guarantee there
are mistakes, so please edit it closely so that we can make sure it's
good.  If you find any mistakes/want to add/remove something, please let
me know as soon as you can so we can make the changes before submitting
it Friday!